requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
dj-database-url==1.2.0
Django==4.1.6
django-heroku==0.3.1
djangorestframework==3.14.0
docopt==0.6.2
gunicorn==20.1.0
idna==3.4
pipreqs==0.4.11
psycopg2==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.3.0
yarg==0.1.9
